WIP switching computers

+++ backend/src/main/java/com/gw/backend/repository/UserPreferencesRepository.java
@@ -24,35 +24,35 @@ public interface UserPreferencesRepository extends JpaRepository<UserPreferences
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artMovement = :artMovement and p.preference = :preference")
-	Double countLikedArtMovementByUserIdAndPreference(@Param("userId") Long userId, @Param("artMovement") String artMovement, @Param("preference") Preference preference);
+	Integer countLikedArtMovementByUserIdAndPreference(@Param("userId") Long userId, @Param("artMovement") String artMovement, @Param("preference") Preference preference);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artYearFinished = :artYearFinished and p.preference = :preference")
-	Double countArtYearFinishedByUserIdAndPreference(@Param("userId") Long userId, @Param("artYearFinished") Integer artYearFinished, @Param("preference") Preference preference);
+	Integer countArtYearFinishedByUserIdAndPreference(@Param("userId") Long userId, @Param("artYearFinished") Integer artYearFinished, @Param("preference") Preference preference);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artType = :artType and p.preference = :preference")
-	Double countArtTypeByUserIdAndPreference(@Param("userId") Long userId, @Param("artType") String artType, @Param("preference") Preference preference);
+	Integer countArtTypeByUserIdAndPreference(@Param("userId") Long userId, @Param("artType") String artType, @Param("preference") Preference preference);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artistId = :artMovement and p.preference = :preference")
-	Double countArtistIdByUserIdAndPreference(@Param("userId") Long userId, @Param("artistId") Long artistId, @Param("preference") Preference preference);
+	Integer countArtistIdByUserIdAndPreference(@Param("userId") Long userId, @Param("artistId") Long artistId, @Param("preference") Preference preference);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artMovement = :artMovement")
-	Double countLikedArtMovementByUserI(@Param("userId") Long userId, @Param("artMovement") String artMovement);
+	Integer countLikedArtMovementByUserI(@Param("userId") Long userId, @Param("artMovement") String artMovement);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artYearFinished = :artYearFinished")
-	Double countArtYearFinishedByUserId(@Param("userId") Long userId, @Param("artYearFinished") Integer artYearFinished);
+	Integer countArtYearFinishedByUserId(@Param("userId") Long userId, @Param("artYearFinished") Integer artYearFinished);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artType = :artType")
-	Double countArtTypeByUserId(@Param("userId") Long userId, @Param("artType") String artType);
+	Integer countArtTypeByUserId(@Param("userId") Long userId, @Param("artType") String artType);
 
 	@Query("select count(p) from user_preference " +
 			"where p.userId = :userId and p.artistId = :artistId")
-	Double countArtistIdByUserId(@Param("userId") Long userId, @Param("artistId") Long artistId);
+	Integer countArtistIdByUserId(@Param("userId") Long userId, @Param("artistId") Long artistId);
