+++ backend/src/main/java/com/gw/backend/service/UserService.java
@@ -14,7 +14,7 @@ public class UserService {
     @Autowired
     UserRepository userRepository;
 
-    public Long getAuthenticatedUserId () {
+    public Integer getAuthenticatedUserId () {
 
     }
